requirements-dev.txt
# requirements-dev.txt - Development Dependencies

# Include base requirements
-r requirements.txt

# ==========================================
# TESTING FRAMEWORKS & TOOLS
# ==========================================
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0        # Parallel test execution
pytest-html>=3.2.0         # HTML test reports
coverage>=7.2.0             # Test coverage
factory-boy>=3.3.0          # Test data factories
faker>=19.3.0               # Fake data generation
model-bakery>=1.12.0        # Django model factories
django-test-plus>=2.2.0     # Enhanced Django testing
freezegun>=1.2.0           # Mock datetime for tests

# ==========================================
# CODE QUALITY & FORMATTING
# ==========================================
black>=23.7.0               # Code formatter
isort>=5.12.0              # Import sorting
flake8>=6.0.0              # Linting
flake8-docstrings>=1.7.0   # Docstring linting
flake8-isort>=6.0.0        # isort integration for flake8
flake8-black>=0.3.0        # black integration for flake8
autoflake>=2.2.0           # Remove unused imports
bandit>=1.7.0              # Security linting

# Type checking
mypy>=1.5.0
django-stubs>=4.2.0
djangorestframework-stubs>=3.14.0
types-requests>=2.31.0
types-redis>=4.6.0

# ==========================================
# PRE-COMMIT & GIT HOOKS
# ==========================================
pre-commit>=3.3.0
commitizen>=3.6.0          # Conventional commits

# ==========================================
# DJANGO DEVELOPMENT TOOLS
# ==========================================
django-extensions>=3.2.0
django-debug-toolbar>=4.1.0
django-silk>=5.0.0         # Profiling and monitoring
django-querycount>=0.8.0   # Query counting middleware

# ==========================================
# API TESTING & DOCUMENTATION
# ==========================================
httpie>=3.2.0             # Command line HTTP client
postman-cli>=1.0.0        # Postman CLI (if available)

# ==========================================
# DATABASE TOOLS
# ==========================================
pgcli>=3.5.0              # PostgreSQL CLI
mycli>=1.27.0             # MySQL CLI (if needed)

# ==========================================
# DOCUMENTATION GENERATION
# ==========================================
sphinx>=7.1.0             # Documentation generator
sphinx-rtd-theme>=1.3.0   # Read the Docs theme
sphinx-autodoc-typehints>=1.24.0
myst-parser>=2.0.0        # Markdown support for Sphinx

# ==========================================
# JUPYTER & NOTEBOOKS
# ==========================================
jupyter>=1.0.0
ipython>=8.14.0
notebook>=6.5.0
jupyterlab>=4.0.0

# Django shell extensions
django-jupyter>=0.0.7
ipdb>=0.13.0              # Debugger

# ==========================================
# PERFORMANCE PROFILING
# ==========================================
line-profiler>=4.1.0
memory-profiler>=0.61.0
py-spy>=0.3.0             # Sampling profiler
django-querycount>=0.8.0

# ==========================================
# SECURITY SCANNING
# ==========================================
safety>=2.3.0             # Check for known security vulnerabilities
pip-audit>=2.6.0          # Audit pip packages for known vulnerabilities

# ==========================================
# DATABASE MIGRATIONS & SCHEMA
# ==========================================
django-migration-linter>=4.1.0  # Lint Django migrations
django-schema-graph>=2.2.0      # Visualize Django models

# ==========================================
# LOCALIZATION & TRANSLATION
# ==========================================
babel>=2.12.0
django-rosetta>=0.9.0     # Translation interface

# ==========================================
# MONITORING & DEBUGGING
# ==========================================
icecream>=2.1.0           # Debugging print statements
rich>=13.4.0              # Rich text and beautiful formatting

# ==========================================
# ENVIRONMENT & CONFIGURATION
# ==========================================
python-dotenv>=1.0.0     # Load environment variables from .env
environs>=9.5.0          # Environment variable parsing

# ==========================================
# BUILD & DEPLOYMENT TOOLS
# ==========================================
wheel>=0.41.0
setuptools>=68.0.0
twine>=4.0.0              # Package uploading
build>=0.10.0             # Build tool

# ==========================================
# LOAD TESTING
# ==========================================
locust>=2.16.0            # Load testing tool

# ==========================================
# STATIC ANALYSIS
# ==========================================
vulture>=2.9.0            # Find dead code
radon>=6.0.0              # Code complexity analysis
xenon>=0.9.0              # Complexity monitoring

# ==========================================
# DEVELOPMENT UTILITIES
# ==========================================
watchdog>=3.0.0          # File system event monitoring
honcho>=1.1.0             # Process manager (like foreman)
invoke>=2.2.0             # Task execution tool

# ==========================================
# DJANGO SPECIFIC DEV TOOLS
# ==========================================
django-livereload-server>=0.4  # Live reload during development
django-reset-migrations>=0.4.0  # Reset migrations utility

# ==========================================
# API CLIENT TESTING
# ==========================================
responses>=0.23.0         # Mock HTTP requests
requests-mock>=1.11.0     # Alternative HTTP mocking
vcrpy>=4.3.0              # Record and replay HTTP interactions

# ==========================================
# DEVELOPMENT WORKFLOW
# ==========================================
tox>=4.6.0                # Testing across environments
nox>=2023.4.0             # Alternative to tox

# ==========================================
# INSTALLATION INSTRUCTIONS
# ==========================================
# To install development dependencies:
# pip install -r requirements-dev.txt
#
# To set up pre-commit hooks:
# pre-commit install
#
# To run tests:
# pytest
#
# To check code quality:
# black .
# isort .
# flake8
# mypy .
#
# To run security checks:
# bandit -r .
# safety check
#
# To generate coverage report:
# pytest --cov=legal_manager --cov-report=html
